Downgraded dotnet version

+++ Overstag/Models/Billing.cs
@@ -11,7 +11,7 @@ namespace Overstag.Models
         public int Amount { get; set; }
         public DateTime Timestamp { get; set; }
         public string EventIDs { get; set; }
-        public byte Payed { get; set; }
+        public int Payed { get; set; }
         public string PayID { get; set; }
     }
 }
